advanced_strategies/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytorch-example"
version = "1.0.0"
description = "Federated Learning with PyTorch and Flower (Advanced Example)"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.19.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "wandb==0.17.8",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "pytorch_example.server_app:app"
clientapp = "pytorch_example.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 10
fraction-fit = 0.5
fraction-evaluate = 0.75
local-epochs = 1
use-wandb = true
wandb-token = ""
use-custom-strategy = true

[tool.flwr.federations]
default = "local-sim"

[tool.flwr.federations.local-sim]
options.num-supernodes = 30

[tool.flwr.federations.researchgrid]
address = "researchgrid.flower.ai"
enable-user-auth = true
options.num-supernodes = 20
